Feature: Changes module creation to use resource name and adds priority
In the modules `GrubKernelConfig`, `KernelModule`, `KernelParameter` and
`SystemdOverrides` there was an issue where the configuration was not
applied correctly if there were multiple NodeConfigs configuring the
previous modules because they were all using the same name for the
configuration file in the node.

Now we use the NodeConfig resource namespaced name as part of the
configuration file name with a new optional `priority` field to set the
lexicographical order of the files.

Signed-off-by: Alejandro Macedo <alex.macedopereira@gmail.com>